refactors zoom level to CGFloat

+++ Metal/RendererViewModel.swift
@@ -12,9 +12,9 @@ import Combine
 protocol MetalbrotViewModelInterface: AnyObject {
     
     //Monstrous @Published Protocol Adapter
-    var zoomLevel: Int { get }
-    var zoomLevelPublished: Published<Int> { get }
-    var zoomLevelPublisher: Published<Int>.Publisher { get }
+    var zoomLevel: CGFloat { get }
+    var zoomLevelPublished: Published<CGFloat> { get }
+    var zoomLevelPublisher: Published<CGFloat>.Publisher { get }
     
     //Monstrous @Published Protocol Adapter
     var center: CGPoint { get }
@@ -22,7 +22,7 @@ protocol MetalbrotViewModelInterface: AnyObject {
     var centerPublisher: Published<CGPoint>.Publisher { get }
     
     func updateCenter(_ newPoint: CGPoint)
-    func updateZoom(_ newZoomLevel: Int)
+    func updateZoom(_ newZoomLevel: CGFloat)
     
     func getAdjustedSize(viewSize: vector_uint2) -> vector_uint2
     func getAdjustedPosition(viewSize: vector_uint2) -> vector_int2
@@ -34,9 +34,9 @@ final class MetalbrotRendererViewModel: MetalbrotViewModelInterface {
 
     
     //MetalbrotViewModelInterface protocol conformance
-    var zoomLevel: Int { zoomLevelConcretePublished }
-    var zoomLevelPublished: Published<Int> { _zoomLevelConcretePublished }
-    var zoomLevelPublisher: Published<Int>.Publisher { $zoomLevelConcretePublished }
+    var zoomLevel: CGFloat { zoomLevelConcretePublished }
+    var zoomLevelPublished: Published<CGFloat> { _zoomLevelConcretePublished }
+    var zoomLevelPublisher: Published<CGFloat>.Publisher { $zoomLevelConcretePublished }
     
     var center: CGPoint { centerConcretePublished }
     var centerPublished: Published<CGPoint> { _centerConcretePublished }
@@ -59,14 +59,14 @@ final class MetalbrotRendererViewModel: MetalbrotViewModelInterface {
         centerConcretePublished = newPoint
     }
     
-    func updateZoom(_ newZoomLevel: Int){
+    func updateZoom(_ newZoomLevel: CGFloat){
         zoomLevelConcretePublished = newZoomLevel
     }
     
     
     
     //Concrete Implementation
-    @Published private var zoomLevelConcretePublished: Int
+    @Published private var zoomLevelConcretePublished: CGFloat
     @Published private var centerConcretePublished: CGPoint
     
     init(){
